Update merging tile group
