+++ .idea/runConfigurations/Serve_Docs_Locally.xml
@@ -1,0 +1,7 @@
+<component name="ProjectRunConfigurationManager">
+  <configuration default="false" name="Serve Docs Locally" type="CompoundRunConfigurationType">
+    <toRun name="samples - run" type="FlutterRunConfigurationType" />
+    <toRun name="docs - run" type="js.build_tools.npm" />
+    <method v="2" />
+  </configuration>
+</component>
