+++ docs/pages/docs/tile/tile-group.mdx
@@ -5,15 +5,11 @@ import LinkBadge from "../../../components/link-badge/link-badge.tsx";
 import LinkBadgeGroup from "../../../components/link-badge/link-badge-group.tsx";
 
 # Tile Group
-A tile group groups multiple tiles.
+A tile group groups multiple groups and tiles.
 
 <LinkBadgeGroup>
     <LinkBadge label="API Reference" href="https://pub.dev/documentation/forui/latest/forui.widgets.tile/forui.widgets.tile-library.html"/>
 </LinkBadgeGroup>
-
-<Callout type="info">
-    Tile groups can be grouped together inside a [FTileGroups](/docs/tile/tile-groups).
-</Callout>
 
 <Tabs items={['Preview', 'Code']}>
   <Tabs.Tab>
@@ -56,9 +52,74 @@ FTileGroup(
 );
 ```
 
+### `FTileGroup.merge(...)`
+
+```dart
+FTileGroup.merge(
+  label: const Text('Settings'),
+  semanticLabel: 'Settings',
+  divider: FTileDivider.indented,
+  children: [],
+);
+```
+
 ## Examples
 
-### Full Divider
+### Merge Multiple Groups
+
+<Tabs items={['Preview', 'Code']}>
+  <Tabs.Tab>
+    <Widget name='tile-group' variant='merge' query={{}} height={300}/>
+  </Tabs.Tab>
+  <Tabs.Tab>
+    ```dart
+    FTileGroup.merge(
+      label: const Text('Settings'),
+      children: [
+        FTileGroup(
+          children: [
+            FTile(
+              prefixIcon: FIcon(FAssets.icons.user),
+              title: const Text('Personalization'),
+              suffixIcon: FIcon(FAssets.icons.chevronRight),
+              onPress: () {},
+            ),
+            FTile(
+              prefixIcon: FIcon(FAssets.icons.wifi),
+              title: const Text('WiFi'),
+              details: const Text('Forus Labs (5G)'),
+              suffixIcon: FIcon(FAssets.icons.chevronRight),
+              onPress: () {},
+            ),
+          ],
+        ),
+        FTileGroup(
+          children: [
+            FTile(
+              prefixIcon: FIcon(FAssets.icons.check),
+              title: const Text('List View'),
+              suffixIcon: FIcon(FAssets.icons.chevronRight),
+              onPress: () {},
+            ),
+            FTile(
+              prefixIcon: const SizedBox.square(
+                dimension: 17,
+              ),
+              title: const Text('Grid View'),
+              suffixIcon: FIcon(FAssets.icons.chevronRight),
+              onPress: () {},
+            ),
+          ],
+        ),
+      ],
+    );
+    ```
+  </Tabs.Tab>
+</Tabs>
+
+### Appearance
+
+#### Full Divider
 
 <Tabs items={['Preview', 'Code']}>
   <Tabs.Tab>
@@ -89,7 +150,7 @@ FTileGroup(
   </Tabs.Tab>
 </Tabs>
 
-### No Divider
+#### No Divider
 
 <Tabs items={['Preview', 'Code']}>
   <Tabs.Tab>
