+++ forui/example/lib/sandbox.dart
@@ -20,52 +20,22 @@ class _SandboxState extends State<Sandbox> {
   }
 
   @override
-  Widget build(BuildContext context) => FTileGroups(
-        label: const Text('Settings'),
-        children: [
-          FTileGroup(
-            label: const Text('Help'),
-            children: [
-              FTile(
-                prefixIcon: FIcon(FAssets.icons.user),
-                title: const Text('Personalization'),
-                suffixIcon: FIcon(FAssets.icons.chevronRight),
-                onPress: () {},
-              ),
-              FTile(
-                prefixIcon: FIcon(FAssets.icons.wifi),
-                title: const Text('WiFi'),
-                details: const Text('Forus Labs (5G)'),
-                suffixIcon: FIcon(FAssets.icons.chevronRight),
-                onPress: () {},
-              ),
-              FTile(
-                prefixIcon: FIcon(FAssets.icons.bell),
-                title: const Text('Notifications'),
-                // subtitle: const Text('Badges, Sound'),
-                suffixIcon: FIcon(FAssets.icons.chevronRight),
-                onPress: () {},
-              ),
-            ],
-          ),
-          FTileGroup(
-            children: [
-              FTile(
-                prefixIcon: FIcon(FAssets.icons.check),
-                title: const Text('List View'),
-                suffixIcon: FIcon(FAssets.icons.chevronRight),
-                onPress: () {},
-              ),
-              FTile(
-                prefixIcon: const SizedBox.square(
-                  dimension: 17,
-                ),
-                title: const Text('Grid View'),
-                suffixIcon: FIcon(FAssets.icons.chevronRight),
-                onPress: () {},
-              ),
-            ],
-          ),
-        ],
-      );
+  Widget build(BuildContext context) => FTileGroup(
+    children: [
+      FTile(
+        prefixIcon: FIcon(FAssets.icons.check),
+        title: const Text('List View'),
+        suffixIcon: FIcon(FAssets.icons.chevronRight),
+        onPress: () {},
+      ),
+      FTile(
+        prefixIcon: const SizedBox.square(
+          dimension: 17,
+        ),
+        title: const Text('Grid View'),
+        suffixIcon: FIcon(FAssets.icons.chevronRight),
+        onPress: () {},
+      ),
+    ],
+  );
 }
