+++ forui/lib/widgets/tile.dart
@@ -8,4 +8,3 @@ library forui.widgets.tile;
 export '../src/widgets/tile/tile.dart' hide extractTile;
 export '../src/widgets/tile/tile_content.dart' hide FTileContent;
 export '../src/widgets/tile/tile_group.dart' hide extractTileGroup;
-export '../src/widgets/tile/tile_groups.dart';
