+++ forui/test/golden/tile/group/merge/empty.png
