+++ forui/test/golden/tile/group/merge/ignore-group-label.png
