+++ forui/test/golden/tile/group/merge/ignore-group-labels.png
