+++ forui/test/golden/tile/group/merge/override.png
