+++ forui/test/golden/tile/group/merge/single.png
