+++ forui/test/golden/tile/group/merge/zinc-dark/FTileDivider.full.png
