+++ forui/test/golden/tile/group/merge/zinc-light/FTileDivider.full.png
