+++ forui/test/golden/tile/group/merge/zinc-light/FTileDivider.indented.png
