+++ forui/test/golden/tile/group/merge/zinc-light/FTileDivider.none.png
